chore: 타입 네이밍 네이밍 변경 및 위치 이동
