+++ src/components/common/Table/CouponProvisionMembersInfoTable.tsx
@@ -3,7 +3,7 @@ import styled from "@emotion/styled";
 import { Button } from "@mui/material";
 import { DataGrid, GridCellParams, GridColDef, GridRowModel } from "@mui/x-data-grid";
 // import useGetIssuedCouponListQuery from "@/hooks/queries/useGetIssuedCouponListQuery";
-import { IssuedCouponType } from "@/types/dtos/coupon";
+import { IssuedCouponType } from "@/types/entities/coupon";
 import { formatDateWithText } from "@/utils/validation/formatDate";
 import { formatPrice } from "@/utils/validation/formatPrice";
 import DetailInfoModal from "../Modal/DetailInfoModal";
